Create README.md
I made a snake game with gui. It contains a directory interface to use actionPerformed to press the button to jump to the game interface, you can open the snake game interface. I first created an interface to draw a static panel using frame. In the game panel I want to create a little snake, and I drew the picture on the static panel. I made the little snake move on the screen and used ActionListener to make the little snake move on the screen. Of course, people need to control it, so I used KeyListener to let the little snake be controlled by the computer and use the space to start or pause the game. Then I drew the initial interface of the game (random location of the food, direction of the snake's head, body coordinates, coordinates of the snake, timer). After that, I made a failure judgment when the snake and body coordinates coincided, and added time to complete it. (Ps: Sometimes the food coordinates overlap with the head, but due to the number of frames, the snake may not be able to eat the food. This is a small flaw). I will also improve this game, such as adding variable speeds, multiple foods , or saving archives and logging in to the interface.